Changed all types to pass-by-reference
This pushes all values out onto the heap, which means more
allocations and more GC load, but a smaller stack and much
less copying of values during function invocations.

This change causes a performance loss of 150ms, so it wasn't the best
performance decision in this case.

Why do it?

* Go docs say "when in doubt, use a pointer"
* It's a better apples-to-apples comparison with all other
  impls, including C
* I was curious if it would make the program faster or slower!